parallelTargets: 2 -> 8 (issue RPT-3193)
in the raw use case, we're fanning out to 12+ targets.  per discussion
in the NB status meeting this morning, there doesn't seem to be any
reason not to move in the direction of having this value be larger
than the number of targets.  however, to be cautious, we've moving in
smaller increments for now; this is the first such increment.

the main worry is that we might suddenly exhaust the memory allocated
to the lambda, and we've recently quadrupled that setting, so that's
the factor we're starting with here as well.